Review/Rating completados

+++ src/main/java/com/dev/videoblogappcore/videoblog/VideoBlogService.java
@@ -44,12 +44,12 @@ public class VideoBlogService {
 
 
 
-    public void createVideoBlog(VideoBlogDTO dto){
+    public void createVideoBlog(VideoBlogDTO dto,String username){
         if(!isRequestCompleted(dto)){
             throw new VideoBlogException(400,"Request no contiene los datos necesarios");
         }
         try{
-            videoBlogRepository.save(dto.toEntity());
+            videoBlogRepository.save(dto.toEntity(username));
         }catch (Exception e)
         {
             throw new VideoBlogException(511,e.getMessage());
@@ -57,11 +57,20 @@ public class VideoBlogService {
     }
 
 
-    public void deleteVideoBlog(String id){
+    public void deleteVideoBlog(String id,String username){
         ObjectId objId = new ObjectId(id);
         Optional<VideoBlog> videoBlogToDelete = videoBlogRepository.findById(objId);
         if(videoBlogToDelete.isPresent()){
-            videoBlogRepository.delete(videoBlogToDelete.get());
+
+            if(videoBlogToDelete.get().getUsername().equals(username)){
+                videoBlogRepository.delete(videoBlogToDelete.get());
+            }else{
+                throw  new VideoBlogException(403,
+                        String.format("El usuario %s no tiene permisos para borrar el video %s. El video blog fue creado por %s",
+                                username,
+                                videoBlogToDelete.get().getTitle(),
+                                videoBlogToDelete.get().getUsername()));
+            }
         }else{
             throw new VideoBlogException(400,"No se encontro un videoBlog con el id :" + id);
         }
